add controller-runtime support
This update adds support to Kubernetes' controller-runtime. v0.15 will
start the controller with controller-runtime by default, but can be
changed to the legacy runtime using command-line option. v0.16 will drop
the legacy runtime support.

All the current runtime features will be migrated, even if in an
incompatible way - and if so, properly documented in the release notes.

Resource parser and configuration builder were marginally changed, just
to fit a few interface adjusts, making this upgrade with a low risk to
generate hidden bugs and compatibility changes.

New architecture at a glance is:

- pkg/controller - continues with the runtime configuration
- pkg/controller/launch - bootstraps all the configurations and
  services needed to start the runtime manager
- pkg/controller/config - static, command-line based config options
- pkg/controller/services - configures, starts and manages all the
  internal services
- pkg/controller/reconciler - watch, filter, group and dispatch
  notification about resource changes
- pkg/controller/legacy - legacy runtime configuration
- pkg/common - legacy runtime component

pkg/controller/legacy and pkg/common will be removed in v0.16.